hotfix

+++ saleor/order/views.py
@@ -112,10 +112,10 @@ def start_payment(request, order, variant):
     """
     order.change_status('payment-pending')
     items = order.get_items()
-    if float(order.total_net.gross) >= settings.DISCOUNT_THRESHOLD:
+    if order.total_net.gross >= int(settings.DISCOUNT_THRESHOLD):
         pagseguro_api = PagSeguroApi(reference='luanavizzon',
                                      sender_email = order.user_email,
-                                     extraAmount = -float(order.total_net.gross)*0.4)
+                                     extraAmount = -order.total_net.gross*0.4)
     elif order.discount_amount is not None:
         pagseguro_api = PagSeguroApi(reference='luanavizzon',
                                      sender_email = order.user_email,
